@@ -34,14 +34,14 @@
     "pre-commit": "node scripts/pre-commit.js && lint-staged"
   },
   "dependencies": {
-    "@angular/animations": "13.3.0",
-    "@angular/common": "13.3.0",
-    "@angular/compiler": "13.3.0",
-    "@angular/core": "13.3.0",
-    "@angular/forms": "13.3.0",
-    "@angular/platform-browser": "13.3.0",
-    "@angular/platform-browser-dynamic": "13.3.0",
-    "@angular/router": "13.3.0",
+    "@angular/animations": "13.3.10",
+    "@angular/common": "13.3.10",
+    "@angular/compiler": "13.3.10",
+    "@angular/core": "13.3.10",
+    "@angular/forms": "13.3.10",
+    "@angular/platform-browser": "13.3.10",
+    "@angular/platform-browser-dynamic": "13.3.10",
+    "@angular/router": "13.3.10",
     "@messageformat/core": "3.0.0",
     "bootstrap": "5.1.0",
     "chalk": "4.1.2",
@@ -59,44 +59,44 @@
     "zone.js": "0.11.5"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "13.3.0",
+    "@angular-devkit/build-angular": "13.3.7",
     "@angular-eslint/eslint-plugin": "13.1.0",
     "@angular-eslint/eslint-plugin-template": "13.1.0",
     "@angular-eslint/template-parser": "13.1.0",
-    "@angular/cli": "13.3.0",
-    "@angular/compiler-cli": "13.3.0",
-    "@angular/language-service": "13.3.0",
+    "@angular/cli": "13.3.7",
+    "@angular/compiler-cli": "13.3.10",
+    "@angular/language-service": "13.3.10",
     "@commitlint/cli": "12.1.4",
     "@commitlint/config-conventional": "12.1.4",
     "@jscutlery/semver": "2.10.0",
     "@ngneat/spectator": "10.0.0",
-    "@nrwl/angular": "13.9.3",
+    "@nrwl/angular": "14.1.6",
     "@nrwl/cli": "13.9.3",
-    "@nrwl/cypress": "13.9.3",
-    "@nrwl/eslint-plugin-nx": "13.9.3",
-    "@nrwl/jest": "13.9.3",
-    "@nrwl/js": "13.9.3",
-    "@nrwl/linter": "13.9.3",
-    "@nrwl/workspace": "13.9.3",
+    "@nrwl/cypress": "14.1.6",
+    "@nrwl/eslint-plugin-nx": "14.1.6",
+    "@nrwl/jest": "14.1.6",
+    "@nrwl/js": "14.1.6",
+    "@nrwl/linter": "14.1.6",
+    "@nrwl/workspace": "14.1.6",
     "@types/command-line-args": "5.2.0",
     "@types/find-duplicated-property-keys": "1.1.0",
     "@types/flat": "5.0.2",
     "@types/fs-extra": "9.0.12",
     "@types/jasmine": "3.8.2",
-    "@types/jest": "27.0.2",
+    "@types/jest": "27.4.1",
     "@types/node": "16.7.2",
-    "@typescript-eslint/eslint-plugin": "5.10.2",
-    "@typescript-eslint/parser": "5.10.2",
+    "@typescript-eslint/eslint-plugin": "5.18.0",
+    "@typescript-eslint/parser": "5.18.0",
     "cypress": "8.7.0",
     "dotenv": "10.0.0",
-    "eslint": "8.7.0",
+    "eslint": "8.12.0",
     "eslint-config-prettier": "8.3.0",
     "eslint-plugin-cypress": "2.11.3",
     "git-cz": "4.7.1",
     "husky": "7.0.2",
     "jasmine-core": "3.10.1",
     "jasmine-spec-reporter": "7.0.0",
-    "jest": "27.2.3",
+    "jest": "27.5.1",
     "jest-preset-angular": "11.1.1",
     "karma": "6.3.17",
     "karma-chrome-launcher": "3.1.0",
@@ -105,17 +105,17 @@
     "karma-jasmine": "4.0.1",
     "karma-jasmine-html-reporter": "1.7.0",
     "lint-staged": "11.1.2",
-    "ng-packagr": "13.2.1",
-    "nx": "13.9.3",
+    "ng-packagr": "13.3.1",
+    "nx": "14.1.6",
     "postcss": "8.3.11",
     "postcss-import": "14.0.2",
     "postcss-preset-env": "6.7.0",
     "postcss-url": "10.1.3",
     "prettier": "2.6.0",
-    "ts-jest": "27.0.5",
+    "ts-jest": "27.1.4",
     "ts-mocks": "3.0.1",
     "ts-node": "10.2.1",
-    "typescript": "4.5.5"
+    "typescript": "4.6.4"
   },
   "config": {
     "commitizen": {
