@@ -1,6 +1,6 @@
-module.exports = {
+export default {
   displayName: 'transloco-schematics',
-  preset: '../../jest.preset.js',
+
   globals: {
     'ts-jest': {
       tsconfig: '<rootDir>/tsconfig.spec.json',
@@ -11,4 +11,5 @@ module.exports = {
   },
   moduleFileExtensions: ['ts', 'tsx', 'js', 'jsx'],
   coverageDirectory: '../../coverage/libs/transloco-schematics',
+  preset: '../../jest.preset.js',
 };
